requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-mdeditor==0.1.20
importlib-metadata==4.11.3
Markdown==3.3.6
mysqlclient==2.1.0
sqlparse==0.4.2
zipp==3.7.0
